delete-events.spec.js file was created for to store the tests of the delete endpoints
